Inclusão do tipo decimal no cálculo de impostos e implementação da função atualizar linha da nota de corretagem

+++ BetterGlauco/parametro.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 from investimento.models import Parametro
 
 class Constante():
@@ -6,13 +7,13 @@ class Constante():
         objeto = Parametro.objects.get_or_create(nome='BDR_CUSTO',
                                                 defaults={'valor': '1.0038',
                                                           'descricao': 'Estimativa de custos operacionais para criar o BDR'})
-        return float(objeto[0].valor)
+        return Decimal(objeto[0].valor)
     
     def BDR_IMPOSTO_DIVIDENDOS_USA(self):
         objeto = Parametro.objects.get_or_create(nome='BDR_IMPOSTO_DIVIDENDOS_USA',
                                                 defaults={'valor': '0.441826',
                                                           'descricao': 'fator para cálculo do imposto de renda sobre dividendos retido nos EUA para os BDRs'})
-        return float(objeto[0].valor)
+        return Decimal(objeto[0].valor)
     
     def E_MAIL_API_SCRAPERLINK(self):
         objeto = Parametro.objects.get_or_create(nome='E_MAIL_API_SCRAPERLINK',
